Add files via upload
I am delighted to share my latest project on GitHub—an engaging Tic-Tac-Toe game built entirely in Python! This project showcases my passion for coding and game development, highlighting my proficiency in utilizing Python functions such as DrawBoard, checkposition, and checkwin to create an immersive gaming experience.

The DrawBoard function allows players to visualize the game board, providing a clear overview of the current state of play. Through careful design, I have made the game interface intuitive and visually appealing, ensuring players have a seamless experience.

To ensure fair gameplay, I implemented the checkposition function, which verifies the validity of each move made by the players. This feature guarantees that players cannot make illegal moves and maintains the integrity of the game.

The checkwin function is the heart of the game, as it determines the winner. By analyzing the board's current configuration, this function intelligently identifies winning conditions, giving players an exhilarating moment of triumph or a friendly reminder to keep trying.